requirements.txt
Flask==0.10.1
Flask-Login==0.2.11
Flask-SQLAlchemy==2.0
Jinja2==2.7.3
MarkupSafe==0.23
MySQL-python==1.2.5
SQLAlchemy==0.9.7
Unidecode==0.04.17
Werkzeug==0.9.6
argparse==1.2.1
boto==2.33.0
google-api-python-client==1.3.1
httplib2==0.9
itsdangerous==0.24
oauth2==1.5.211
oauth2client==1.4.6
pyasn1==0.1.7
pyasn1-modules==0.0.5
python-dateutil==2.0
python-rdio==0.6
python-slugify==0.1.0
rsa==3.1.4
simplejson==3.6.5
six==1.9.0
uritemplate==0.6
wsgiref==0.1.2
